move MetadataAssertions to utils
