+++ app/controllers/DatasetController.scala
@@ -25,7 +25,7 @@ import play.api.libs.json._
 import play.api.mvc.{Action, AnyContent, PlayBodyParsers}
 import play.silhouette.api.Silhouette
 import security.{URLSharing, WkEnv}
-import utils.{ObjectId, WkConf}
+import utils.{MetadataAssertions, ObjectId, WkConf}
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
@@ -67,14 +67,6 @@ case class SegmentAnythingMaskParameters(
 
 object SegmentAnythingMaskParameters {
   implicit val jsonFormat: Format[SegmentAnythingMaskParameters] = Json.format[SegmentAnythingMaskParameters]
-}
-
-trait MetadataAssertions {
-  def assertNoDuplicateMetadataKeys(metadata: JsArray)(implicit ec: ExecutionContext,
-                                                       provider: MessagesProvider): Fox[Unit] = {
-    val keys = metadata.value.flatMap(_.as[JsObject] \\ "key").map(_.as[String]).toList
-    if (keys.size == keys.distinct.size) Fox.successful(()) else Fox.failure(Messages("dataset.metadata.duplicateKeys"))
-  }
 }
 
 class DatasetController @Inject()(userService: UserService,
